+++ pallets/tokens/src/currencies.rs
@@ -14,154 +14,10 @@
  * limitations under the License.
  */
 
-use crate::*;
+use crate::{mutations::Mutation, Module, RawEvent, Trait};
 use frame_support::traits::BalanceStatus;
 use governance_os_support::traits::{Currencies, ReservableCurrencies};
-use sp_runtime::{
-    traits::{CheckedAdd, CheckedSub, Saturating},
-    DispatchError, DispatchResult,
-};
-use sp_std::{collections::btree_map::BTreeMap, marker};
-
-/// An internal helper to represent balance changes. It is used to express in a better manner
-/// operations done on balances while saving on weight costs by fetching the required data
-/// only when necessary and forwarding errors approprietaly.
-#[derive(Clone)]
-struct Mutation<T: Trait> {
-    currency_id: T::CurrencyId,
-    balances: BTreeMap<T::AccountId, (AccountCurrencyData<T::Balance>, bool)>,
-    coins_created: T::Balance,
-    coins_burned: T::Balance,
-    _phantom: marker::PhantomData<T>,
-}
-impl<T: Trait> Mutation<T> {
-    fn new_for_currency(currency_id: T::CurrencyId) -> Self {
-        Self {
-            currency_id: currency_id,
-            balances: BTreeMap::new(),
-            coins_created: 0.into(),
-            coins_burned: 0.into(),
-            _phantom: marker::PhantomData,
-        }
-    }
-
-    /// Return a balance from the `self.balances` cache or fetch it from the node DB.
-    fn get_or_fetch_balance(&mut self, who: &T::AccountId) -> AccountCurrencyData<T::Balance> {
-        let in_cache = self.balances.get(who);
-        match in_cache {
-            None => {
-                let in_node = Balances::<T>::get(who, self.currency_id);
-                self.balances.insert(who.clone(), (in_node.clone(), false));
-
-                in_node
-            }
-            Some(cache) => cache.clone().0,
-        }
-    }
-
-    /// Save an updated balance in our memory cache
-    fn save_balance(&mut self, who: &T::AccountId, balance: AccountCurrencyData<T::Balance>) {
-        // Note how the boolean is set to true since we modified the balance
-        self.balances.insert(who.clone(), (balance, true));
-    }
-
-    /// Verify that the currency is transferable
-    fn ensure_must_be_transferable_for(&mut self, who: &T::AccountId) -> DispatchResult {
-        if !RoleManagerOf::<T>::has_role(
-            who,
-            RoleBuilderOf::<T>::transfer_currency(self.currency_id),
-        ) {
-            return Err(Error::<T>::UnTransferableCurrency.into());
-        }
-
-        Ok(())
-    }
-
-    fn add_free_balance(&mut self, who: &T::AccountId, increment: T::Balance) -> DispatchResult {
-        let mut balance = self.get_or_fetch_balance(who);
-        balance.free = balance
-            .free
-            .checked_add(&increment)
-            .ok_or(Error::<T>::BalanceOverflow)?;
-        self.coins_created = self.coins_created.saturating_add(increment);
-        self.save_balance(who, balance);
-
-        Ok(())
-    }
-
-    fn sub_free_balance(&mut self, who: &T::AccountId, decrement: T::Balance) -> DispatchResult {
-        let mut balance = self.get_or_fetch_balance(who);
-        balance.free = balance
-            .free
-            .checked_sub(&decrement)
-            .ok_or(Error::<T>::BalanceTooLow)?;
-        self.coins_burned = self.coins_burned.saturating_add(decrement);
-        self.save_balance(who, balance);
-
-        Ok(())
-    }
-
-    fn add_reserved_balance(
-        &mut self,
-        who: &T::AccountId,
-        increment: T::Balance,
-    ) -> DispatchResult {
-        let mut balance = self.get_or_fetch_balance(who);
-        balance.reserved = balance
-            .reserved
-            .checked_add(&increment)
-            .ok_or(Error::<T>::BalanceOverflow)?;
-        self.coins_created = self.coins_created.saturating_add(increment);
-        self.save_balance(who, balance);
-
-        Ok(())
-    }
-
-    fn sub_up_to_reserved_balance(
-        &mut self,
-        who: &T::AccountId,
-        decrement: T::Balance,
-    ) -> T::Balance {
-        let mut balance = self.get_or_fetch_balance(who);
-        let actual_subed = balance.reserved.min(decrement);
-        // We just capped `actual_subed` to `balance.reserved` itself.
-        balance.reserved -= actual_subed;
-        self.coins_burned = self.coins_burned.saturating_add(actual_subed);
-        self.save_balance(who, balance);
-
-        actual_subed
-    }
-
-    /// Commit all the changes to the chain state or error.
-    fn apply(self) -> DispatchResult {
-        self.balances
-            .iter()
-            .filter(|(_account, (_bal, changed))| *changed)
-            .map(|(account, (balance, _changed))| (account, balance))
-            .for_each(|(account, balance)| {
-                if balance.total() == Zero::zero() {
-                    Balances::<T>::remove(account, self.currency_id);
-                } else {
-                    Balances::<T>::insert(account, self.currency_id, balance);
-                }
-            });
-
-        if self.coins_created != self.coins_burned {
-            TotalIssuances::<T>::try_mutate(self.currency_id, |d| -> DispatchResult {
-                *d = d
-                    .checked_sub(&self.coins_burned)
-                    .ok_or(Error::<T>::TotalIssuanceUnderflow)?;
-                *d = d
-                    .checked_add(&self.coins_created)
-                    .ok_or(Error::<T>::TotalIssuanceOverflow)?;
-
-                Ok(())
-            })?;
-        }
-
-        Ok(())
-    }
-}
+use sp_runtime::{traits::Saturating, DispatchError, DispatchResult};
 
 impl<T: Trait> Currencies<T::AccountId> for Module<T> {
     type CurrencyId = T::CurrencyId;
